캡슐화
