+++ TIL/topics/java_basic/2025.06.17 .md
@@ -2,9 +2,18 @@
 # TIL(Today I Learned)
 
 ## Test Code
- - assertEquals();
-
- - assertTrue();
+ - assertEquals(); / assertNotEquals();
+    두 값이 같은지/다른지 확인
+ - assertSame(); / assertNotSame();
+    두 객체의 참조가 같은지/다른지 확인
+ - assertTrue(); / assertFalse();
+    조건이 true/false 인지 확인
+ - assertNull(); / assertNotNull()
+   객체가 null이 인지/아닌지 확인
+ - assertArrayEquals();
+   두 배열이 같은지 확인
+ - assertThrows(Exception.class, () -> { ... });
+   예외 발생 여부 확인
 
 ## commit 메세지
  - feat: 새로운 기능 추가
@@ -18,17 +27,16 @@
  - ci: CI 설정 파일 수정
  - perf: 성능 개선
  - chore: 빌드 업무 수정, 패키지 매니저 수정
- - rename: 파일 혹은 풀더명을 수정만 한 경우
+ - rename: 파일 혹은 풀더 명을 수정만 한 경우
  - remove: 파일을 삭제만 한 경우
 
 
-객체지향의 3대 원칙
-캡슐화 (encapsulation)
-상속
-다양성
-(추상화)
+## 객체 지향의 3대 원칙 (4대의 경우 추상화 포함)
+ - 캡슐화 (encapsulation)
+ - 상속
+ - 다양성
 
-오버로드
+오버 로드
 없앨 수 없다.
 안전한 클래스를 만들어야 하는데!!
 
@@ -46,5 +54,25 @@ getter와 setter
 getter - 읽기 전용
 setter - 쓰기 전용
 
-보일러플레이트
+보일러 플레이트
+
+## 자료 구조
+동적 배열(Array)
+
+연결 리스트(List)
+ - 단순 연결 리스트
+
+스택(Stack)
+
+큐(Queue)
+
+## 컬렉션
+List - 
+Map - 
+Set - 
+
+Iterator<String> it = names.iterator;
+
+HashSet - 컨테인즈
+HashMap - 빠르다
