--- src/test/java/com/survivalcoding/assignments_01_instance/exam01/KinokoTest.java
@@ -1,25 +1,0 @@
-package com.survivalcoding.assignments_01_instance.exam01;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class KinokoTest {
-
-    /*@Test //옵션 설명이 같이 나온다?
-    @DisplayName("heal을 하면 hp를 10 회복해야 한다")
-    void heal() {
-        // given(준비)
-        final Wizard wizard = new Wizard("마법사", 100);
-        final Hero hero = new Hero("히어로", 10);
-
-        // when(실행)
-        wizard.heal(hero);
-
-        // then(검증)
-        assertEquals(20, hero.hp);
-    }*/
-
-
-}
